Add useUpdate hooks
